feat editorCam : full input implementation
this commit should:

implement classes for mouse move result plus key results for press/release.
allow the editor cam to listen for the keys and notify the resulting code to execute on notify.
hide cursor on right click
show cursor on right click release
rotate editor cam on mouse move as long as right click is currently on the lock state.